refactor;

+++ MeetingRooms-iOS/Managers/GoogleAccountManager/MRNetworkManager.m
@@ -118,7 +118,8 @@ NSString *const baseURL = @"http://redmine.cleveroad.com:3503";
     NSDateFormatter *formatter = [NSDateFormatter new];
     formatter.dateFormat = @"YYYY-MM-dd";
     NSString *dateString = [formatter stringFromDate:date];
-    NSString *tempString = [NSString stringWithFormat:@"%@/api/v1/users/%ld/bookings?date=%@&offset=%lu", baseURL, (long)self.owner.userId.integerValue, dateString, offset];    [self. manager GET:tempString parameters:nil progress:nil success:^(NSURLSessionDataTask * _Nonnull task, id  _Nullable responseObject) {
+    NSString *tempString = [NSString stringWithFormat:@"%@/api/v1/users/%ld/bookings?date=%@&offset=%lu", baseURL, (long)self.owner.userId.integerValue, dateString, (long)offset];
+    [self.manager GET:tempString parameters:nil progress:nil success:^(NSURLSessionDataTask * _Nonnull task, id  _Nullable responseObject) {
         if ([responseObject isKindOfClass:[NSDictionary class]]) {
             NSArray *bookings =  [responseObject valueForKey:@"bookings"];
             NSMutableArray *arrayOfMeetings = [NSMutableArray new];
